fix clippy
