--- crates/iota-genesis-builder/src/stardust/types/address_swap_map/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (c) 2024 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod address_swap_map;
-mod tests;
-pub use address_swap_map::AddressSwapMap;
